% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.resid.R
\name{get.resid}
\alias{get.resid}
\title{get.resid}
\usage{
get.resid(X, Y, A)
}
\arguments{
\item{X}{n x p feature matrix}

\item{Y}{vector specifying outcome variable of interest (e.g., disease status)}

\item{A}{matrix of nuisance variables}
}
\value{
list containing X.tilde (goes in penalty term for PeDecURe) and X.star (goes in maximization term for PeDecURe)
}
\description{
Function to get residuals needed for PeDecURe
}
