% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.cor.R
\name{partial.cor}
\alias{partial.cor}
\title{partial correlation}
\usage{
partial.cor(scores, A, Y)
}
\arguments{
\item{scores}{matrix of PC scores (e.g., Xv1, Xv2, Xv3)}

\item{A}{matrix of nuisance variables}

\item{Y}{outcome variable}
}
\value{
list of partial and marginal correlations between each PC score and A and Y
}
\description{
estimate partial (+ marginal) correlations between each PC score and Y|A and A|Y
}
