% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedecure.R
\name{pedecure}
\alias{pedecure}
\title{PeDecURe}
\usage{
pedecure(
  X,
  X.penalize = NULL,
  A,
  Y,
  lambda,
  nPC,
  centerX = F,
  centerA = F,
  centerY = F,
  scaleX = F,
  scaleA = F,
  scaleY = F
)
}
\arguments{
\item{X}{n x p matrix of features to appear in the PeDecURe maximization term. In PeDecURe, this should be defined as X* ("X star"), i.e. the matrix of residuals after subtracting out the effects of A conditional on Y}

\item{X.penalize}{n x p matrix of features to appear in the penalty term.In PeDecURe, this should be defined as X~ ("X tilde"), i.e. the matrix of residuals after subtracting out the effects of Y conditional on A. If X.penalize is not specified, no penalization will be done.}

\item{A}{matrix of nuisance variables}

\item{Y}{vector specifying outcome variable of interest (e.g., disease status)}

\item{lambda}{tuning parameter (to be determined in other function)}

\item{nPC}{number of primary components (PCs) to be extracted}

\item{centerX}{TRUE/FALSE whether to center X by column mean (note: some centering/scaling may have been done before inputting data into pedecure function)}

\item{centerA}{TRUE/FALSE whether to center A by column mean}

\item{centerY}{TRUE/FALSE whether to center Y by mean}

\item{scaleX}{TRUE/FALSE whether to scale each column of X by its standard deviation}

\item{scaleA}{TRUE/FALSE whether to scale each column of A by its standard deviation}

\item{scaleY}{TRUE/FALSE whether to scale Y by its standard deviation}
}
\value{
PC loading vectors
}
\description{
Function to implement PeDecURe (or PLS or other penalized PLS method, depending on input specifications). Note: the code for implementing PeDecURe was adapted from Lin et al. (2016)'s implementation of AC-PCA in R (see https://github.com/linzx06/AC-PCA).
}
