% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedecure.tune.R
\name{pedecure.tune}
\alias{pedecure.tune}
\title{pedecure.tune}
\usage{
pedecure.tune(
  X.orig,
  X.max,
  X.penalize,
  lambdas,
  A,
  Y,
  nPC,
  centerX = F,
  scaleX = F,
  centerY = F,
  scaleY = F,
  centerA = F,
  scaleA = F,
  n.cores = 1,
  plot = FALSE,
  thresh = 0.001
)
}
\arguments{
\item{X.orig}{n x p matrix of ~original~ features (not residualized). this is the version of the data that's used to obtain PC scores}

\item{X.max}{n x p matrix of features that goes into the maximization term (X* ("X star") for PeDecURe)}

\item{X.penalize}{n x p matrix of features to appear in the penalty term.In PeDecURe, this should be defined as X~ ("X tilde"), i.e. the matrix of residuals after subtracting out the effects of Y conditional on A.}

\item{lambdas}{vector including candidate values of tuning parameter (e.g., lambdas = seq(0,10,by=0.1))}

\item{A}{matrix of nuisance variables}

\item{Y}{vector specifying outcome variable of interest (e.g., disease status)}

\item{nPC}{number of primary components (PCs) to be extracted}

\item{centerX}{TRUE/FALSE whether to center X by column mean (note: some centering/scaling may have been done before inputting data into pedecure function)}

\item{scaleX}{TRUE/FALSE whether to scale each column of X by its standard deviation}

\item{centerY}{TRUE/FALSE whether to center Y by mean}

\item{scaleY}{TRUE/FALSE whether to scale Y by its standard deviation}

\item{centerA}{TRUE/FALSE whether to center A by column mean}

\item{scaleA}{TRUE/FALSE whether to scale each column of A by its standard deviation}

\item{n.cores}{number of cores for parallelization in parallel::mclapply()}

\item{plot}{TRUE/FALSE whether to plot the tuning function for different values of lambda}

\item{thresh}{threshold for deciding when lambda is good enough (unit is proportion change, not percentage)}
}
\value{
best lambda
}
\description{
Function to tune lambda for PeDecURe. note: input arguments are similar but not identical to those in pedecure() function
}
