% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedecureMax.R
\name{pedecureMax}
\alias{pedecureMax}
\title{pedecureMax}
\usage{
pedecureMax(v, args)
}
\arguments{
\item{v}{PC loading vector that will be maximized via eigs_sym}

\item{args}{list of arguments (inputted from within pedecure() function)}
}
\value{
v which maximizes objective function
}
\description{
maximization function for PeDecURe (called within function pedecure()). Note: the code for implementing PeDecURe was adapted from Lin et al. (2016)'s implementation of AC-PCA in R (see https://github.com/linzx06/AC-PCA).
}
