% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsMax.R
\name{plsMax}
\alias{plsMax}
\title{plsMax}
\usage{
plsMax(v, args)
}
\arguments{
\item{v}{PC loading vector that will be maximized via eigs_sym}

\item{args}{list of arguments (inputted from within pedecure() function)}
}
\value{
v which maximizes v'X'YY'Xv
}
\description{
maximization function for PLS, when no penalty term is specified in pedecure() function
}
